polkadot-collator: Switch to wasm only (#1054)
* polkadot-collator: Switch to wasm only

This switches the polkadot-collator to run everything in wasm only mode. While we should not that
yet with the relay chain, because it can happen that we run out of memory (very unlikely). On the
relay chain that would be bad, because we only have at max 2 sessions to bring everything back, for
Parachains that isn't such a problem as they would only stall and we could roll out a release that
fixes it. Besides that, Parachain validation on the relay chain happens in Wasm already all the time
and there is the memory usage even higher then on block import.

* cargo fmt

* remove unused var

Co-authored-by: Squirrel <gilescope@gmail.com>